+++ frontend/src/components/interactive/RoundedContainerButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import styled from "styled-components";
-import { DropdownTooltip } from "./DropdownTooltip";
+import { DropdownTooltip } from "../dropdowns/DropdownTooltip";
 import {ReactComponent as RoundedContainerIcon} from '../svg/RoundedContainerIcon.svg';
 
 export const RoundedContainerButton = (props: RoundedContainerButtonProps) => {
