+++ frontend/src/components/interactive/SlideCheckbox.tsx
